cleaning b-baner for captcha
