+++ _src-lp/blocks/banner-with-cards/banner-with-cards.js
@@ -1,4 +1,3 @@
-import { activateRadios } from '../../scripts/delayed.js';
 import { productAliases } from '../../scripts/scripts.js';
 import { updateProductsList } from '../../scripts/utils.js';
 
@@ -62,6 +61,36 @@ function createBuyButtons(tableBuyBtn, prodName, onSelectorClass, onSelectorClas
   const buyButtonMonthly = createButton(onSelectorClassM, 'monthly show');
 
   tableBuyBtn.innerHTML = `${buyButtonYearly} ${buyButtonMonthly}`;
+}
+
+function activateRadios(block, type) {
+  const allRadios = block.querySelectorAll('input[type="radio"]');
+  allRadios.forEach((radio) => {
+    const radioID = radio.id;
+    block.addEventListener('click', (event) => {
+      const { target } = event;
+      if (target.tagName === 'INPUT' && target.id === radioID) {
+        const select = target.dataset.select;
+        let prodBox = target.closest('.prodBox');
+
+        if (typeof type !== 'undefined' && type === 'combined') {
+          prodBox = event.target.closest('.hasProds');
+          allRadios.forEach((r) => {
+            r.checked = (r.dataset.select === select);
+          });
+        }
+
+        if (prodBox) {
+          ['yearly', 'monthly'].forEach((period) => {
+            prodBox.querySelectorAll(`.${period}`).forEach((item) => {
+              item.classList.toggle('show');
+              item.classList.toggle('hide');
+            });
+          });
+        }
+      }
+    });
+  });
 }
 
 let counter = 0;
@@ -160,5 +189,7 @@ export default function decorate(block) {
   `;
 
   // activate radios
-  activateRadios(block, type);
+  setTimeout(function() {
+    activateRadios(block, type);
+  }, 2000);
 }
